+++ CHANGELOG.md
@@ -9,6 +9,7 @@ You can also check [on GitHub](https://github.com/nextcloud/news/releases), the 
 - Added a feature to download the web version of an article on demand, to avoid downloading unnecessary content
 
 ### Fixed
+- Show a user-facing error when moving a feed fails (#3649)
 
 # Releases
 ## [28.2.0-beta.1] - 2026-03-22